credit_approval_project/requirements.txt
Django>=4.0,<5.0
djangorestframework
psycopg2-binary
celery
redis
pandas
python-dotenv
openpyxl